+++ packages/integration/src/addFileScope.ts
@@ -1,5 +1,11 @@
-import { posix, relative, sep } from 'path';
+import { posix, relative, win32 } from 'path';
 import { detectSyntax } from 'mlly';
+import outdent from 'outdent';
+
+// Inlined from @rollup/pluginutils
+// https://github.com/rollup/plugins/blob/33174f956304ab4aad4bbaba656f627c31679dc5/packages/pluginutils/src/normalizePath.ts#L5-L7
+export const normalizePath = (filename: string) =>
+  filename.split(win32.sep).join(posix.sep);
 
 interface AddFileScopeParams {
   source: string;
@@ -16,54 +22,46 @@ export function addFileScope({
   globalAdapterIdentifier,
 }: AddFileScopeParams) {
   // Encode windows file paths as posix
-  const normalizedPath = posix.join(...relative(rootPath, filePath).split(sep));
-
-  if (source.indexOf('@vanilla-extract/css/fileScope') > -1) {
-    return source.replace(
-      /setFileScope\(((\n|.)*?)\)/,
-      `setFileScope("${normalizedPath}", "${packageName}")`,
-    );
-  }
+  const normalizedPath = normalizePath(relative(rootPath, filePath));
 
   const { hasESM, isMixed } = detectSyntax(source);
 
-  if (hasESM && !isMixed) {
-    const setAdapterSnippet = globalAdapterIdentifier
-      ? `
-        import * as __vanilla_css_adapter__ from "@vanilla-extract/css/adapter";
-        __vanilla_css_adapter__.setAdapter(${globalAdapterIdentifier});
-      `
-      : '';
-    const removeAdapterSnippet = setAdapterSnippet
-      ? '__vanilla_css_adapter__.removeAdapter();'
-      : '';
+  if (source.includes('@vanilla-extract/css/fileScope')) {
+    source = source.replace(
+      /setFileScope\(((\n|.)*?)\)/,
+      `setFileScope("${normalizedPath}", "${packageName}")`,
+    );
+  } else {
+    if (hasESM && !isMixed) {
+      source = outdent`
+        import { setFileScope, endFileScope } from "@vanilla-extract/css/fileScope";
+        setFileScope("${normalizedPath}", "${packageName}");
+        ${source}
+        endFileScope();
+      `;
+    } else {
+      source = outdent`
+        const __vanilla_filescope__ = require("@vanilla-extract/css/fileScope");
+        __vanilla_filescope__.setFileScope("${normalizedPath}", "${packageName}");
+        ${source}
+        __vanilla_filescope__.endFileScope();
+      `;
+    }
+  }
 
-    return `
-      ${setAdapterSnippet}
-      import { setFileScope, endFileScope } from "@vanilla-extract/css/fileScope";
-      setFileScope("${normalizedPath}", "${packageName}");
+  if (globalAdapterIdentifier) {
+    const adapterImport =
+      hasESM && !isMixed
+        ? 'import * as __vanilla_css_adapter__ from "@vanilla-extract/css/adapter";'
+        : 'const __vanilla_css_adapter__ = require("@vanilla-extract/css/adapter");';
+
+    source = outdent`
+      ${adapterImport}
+      __vanilla_css_adapter__.setAdapter(${globalAdapterIdentifier});
       ${source}
-      endFileScope();
-      ${removeAdapterSnippet}
+      __vanilla_css_adapter__.removeAdapter();
     `;
   }
 
-  const setAdapterSnippet = globalAdapterIdentifier
-    ? `
-      const __vanilla_css_adapter__ = require("@vanilla-extract/css/adapter");
-      __vanilla_css_adapter__.setAdapter(${globalAdapterIdentifier});
-    `
-    : '';
-  const removeAdapterSnippet = setAdapterSnippet
-    ? '__vanilla_css_adapter__.removeAdapter();'
-    : '';
-
-  return `
-    ${setAdapterSnippet}
-    const __vanilla_filescope__ = require("@vanilla-extract/css/fileScope");
-    __vanilla_filescope__.setFileScope("${normalizedPath}", "${packageName}");
-    ${source}
-    __vanilla_filescope__.endFileScope();
-    ${removeAdapterSnippet};
-  `;
+  return source;
 }