+++ fixtures/layers/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @fixtures/layers
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [[`f373d7f`](https://github.com/vanilla-extract-css/vanilla-extract/commit/f373d7f6b59f43236dc713e1b421ef4631f392c0)]:
+  - @vanilla-extract/css@1.14.1
+
 ## 0.0.6
 
 ### Patch Changes